SystemUI: option to toggle dimming of the notification shade (1/2)
Allow the user to enable/disable dimming of the notification shade
pulldown. Check isHighEndGfx() to respect the default setting that
the device currently supports.

Users on low-end device may want to have this feature, while users
on high-end devices may want to disable it (possible battery life
savings and what not).

Change-Id: I730c0eea266541468d2d38ef88e136d85e1e4bfa

Conflicts:
	core/java/android/provider/Settings.java