FW : Allow user selection of Font Size
ps 1:  This re-structures a lot of the way icons & layouts are sized.
It allows System Icons, Notification Icons, etc. to scale according
to the size of the statusbar.  StatusBar now sizes itself based on
the height of the Font (which is user selectable in RC).

This needs a lot of testing, as I've made substantial changes to the
way the Statusbar.  Also, While everything is 'working', the alignment
of the Textviews alludes me.  All textviews are being aligned at the bottom
of the StatusBar rather than centering like they should.

Change-Id: I5f9e99930210d3d91cda3ef167ad42993c261816
Signed-off-by: Zaphod <mwilson@acesappraisal.com>

Conflicts:
	packages/SystemUI/res/layout/battery_cluster_view.xml
	packages/SystemUI/res/layout/status_bar.xml
	packages/SystemUI/src/com/android/systemui/statusbar/SignalClusterView.java
	packages/SystemUI/src/com/android/systemui/statusbar/StatusBarIconView.java
	packages/SystemUI/src/com/android/systemui/statusbar/policy/Clock.java
	policy/src/com/android/internal/policy/impl/PhoneWindowManager.java