Frameworks: SlimPIE the second (1/2)
This commit changes the following:
- add possibility to set trigger height on the left and right trigger
- add possibility to set gravity on the lef and right trigger
- change touch detection logic and calculation (it shows now up immediately...super fast :p)
- reduces values in dimens.xml because of the new touch detection much lesser are needed
- remove on settings the shown values on the left @seekbar on entries where it does not make sense
- reset option for trigger dimensions
- minor string changes
- cleanup
- minor bugfixes

for me it is now like a PIE should be lol

IMPORTANT FOR ALL DM'S:
I reduced the dimens.xml to values which are for most of our phones good
to go as default value (tuna, crespo, nozomi, grouper can handle even lower
values (3dp) which the user can set of course). d2 and other should be fine as well
with the default value. sgs I series I think need an overlay to push them
a bit up. Please test and submit your overlays to our gerrit.

cheers

Change-Id: I4434f54c0ef858b239c112e8071b677d3bb519d8

Conflicts:
	core/java/android/provider/Settings.java