Frameworks: SlimPIE trigger corrections
Sometimes it is good to release stuff and let the ppl test...so some
smaller issues on the automatic trigger detection for navbar and statusbar
came up and we can resolve the last one. This commit addresses following
reported issues:

- Fix bottom trigger show up on landscape when navbar is at the right side
- Fix if expanded desktop mode is "hide only navbar"...statusbar trigger was on
- Fix respect navbar 0dp values you can set for navbar
- a bit comments in the code :p

For the guys who want to see how the triggers are reacting...turn in
Basestatusbar debug value on....and you can permanently see what
the triggers are doing. Tried now all conditions which are possible
as well on tablet....cannot find any issue now.

cheers

Change-Id: I77698c55af2db5d35a2385761b67cb000a8a5b55

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/BaseStatusBar.java

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/BaseStatusBar.java