Give floating flag to previous activities in the stack
When launching stacked activities, only the last one is flagged as floating in sysui when tapping halo.
This causes the system to show the launcher when the first stacked activity is launched (this first intent gets the flag FLAG_ACTIVITY_TASK_ON_HOME
in TaskStackBuilder).
This change makes all the stacked activities get the floating flag if the last one has it.

Change-Id: Ia3d2a06c92bd1410585fa8c46364915405bd568b
Signed-off-by: martincz <martincz@mokeetek.com>