PhoneStatusBar: fix massive surface redraws due of animation race
the wifi/carrier label is animated from AOSP via an alpha animation. Problem
is that we quiet often call this on other animations and this will end in a
animation loop which forces massive surface redraws. Whatever to avoid this
do it like for notification shortcuts and set directly the alpha value which
look so or so better in the whole animation.

As well a small code cleanup

Change-Id: If620db28d9c38ce3adeeef59fdf65c6394d45f11

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java