base: allow to choose CRT animation by default
Since Android 4.4 Electronbean is the default screen-off-animation.
<bool name="config_animateScreenLights">true</bool> seems not to work
as it should at the moment. It will only disable the option to choose
a different screen-off-animation. It will not force to fade-off,
set it to "false" will add the option inside Settings -> Display
to choose between the different animation. At the moment we need
<bool name="config_animateScreenLights">false</bool>
inside the device overlay to make that option visible.

Change-Id: I5c341ae79efddbafb7996dfc1239cbc0b5aadab2